requirements.txt
Pillow==10.0.1
build==1.0.3
twine==4.0.2
